requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
docutils==0.21.2
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pages==0.3
pandas==2.2.2
piglet-templates==1.3.1
pluggy==1.5.0
pycparser==2.22
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
selenium==4.20.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.38.4
wsproto==1.2.0
zipp==3.18.1
